+++ sokannonser/rest/model/queries.py
@@ -20,6 +20,8 @@ swagger_doc_params = {
     settings.X_FEATURE_DISABLE_SMART_FREETEXT: "Disables machine learning enriched queries."
     " Freetext becomes traditional freetext query according to the setting of " 
     "\"%s\"" % settings.X_FEATURE_FREETEXT_BOOL_METHOD,
+    settings.X_FEATURE_ENABLE_FALSE_NEGATIVE: "Enables extra search for the current known "
+    "term in free text to avoid false negatives ",
     settings.PUBLISHED_AFTER: "Fetch job ads published after specified date and time."
     "Accepts either datetime (format YYYY-mm-ddTHH:MM:SS) or number of minutes "
     "(e.g 120 means published in the last two hours)",
@@ -107,7 +109,9 @@ base_annons_query.add_argument(settings.X_FEATURE_FREETEXT_BOOL_METHOD, choices=
                                default=settings.DEFAULT_FREETEXT_BOOL_METHOD,
                                location='headers', required=False)
 base_annons_query.add_argument(settings.X_FEATURE_DISABLE_SMART_FREETEXT, type=inputs.boolean,
-                               location='headers', required=False)
+                               location='headers', required=False),
+base_annons_query.add_argument(settings.X_FEATURE_ENABLE_FALSE_NEGATIVE, type=inputs.boolean,
+                               location='headers', required=False),
 base_annons_query.add_argument(settings.PUBLISHED_BEFORE,
                                type=lambda x: datetime.strptime(x, '%Y-%m-%dT%H:%M:%S'))
 # annons_complete_query.add_argument(settings.PUBLISHED_AFTER,